requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==5.0.2
django-crispy-forms==2.1
sqlparse==0.4.4
tzdata==2024.1
pillow~=10.2.0
